NKAI: port exploration from VCAI
