+++ AI/Nullkiller/Behaviors/ExplorationBehavior.h
@@ -1,0 +1,38 @@
+/*
+* ExplorationBehavior.h, part of VCMI engine
+*
+* Authors: listed in file AUTHORS in main folder
+*
+* License: GNU General Public License v2.0 or later
+* Full text of license available in license.txt file, in main folder
+*
+*/
+#pragma once
+
+#include "lib/VCMI_Lib.h"
+#include "../Goals/CGoal.h"
+#include "../AIUtility.h"
+
+namespace NKAI
+{
+namespace Goals
+{
+	class ExplorationBehavior : public CGoal<ExplorationBehavior>
+	{
+	public:
+		ExplorationBehavior()
+			:CGoal(EXPLORATION_BEHAVIOR)
+		{
+		}
+
+		TGoalVec decompose(const Nullkiller * ai) const override;
+		std::string toString() const override;
+
+		bool operator==(const ExplorationBehavior & other) const override
+		{
+			return true;
+		}
+	};
+}
+
+}
