+++ AI/Nullkiller/CMakeLists.txt
@@ -15,6 +15,8 @@ set(Nullkiller_SRCS
 		Pathfinding/Rules/AIMovementToDestinationRule.cpp
 		Pathfinding/Rules/AIPreviousNodeRule.cpp
 		Pathfinding/ObjectGraph.cpp
+		Pathfinding/GraphPaths.cpp
+		Pathfinding/ObjectGraphCalculator.cpp
 		AIUtility.cpp
 		Analyzers/ArmyManager.cpp
 		Analyzers/HeroManager.cpp
@@ -42,6 +44,7 @@ set(Nullkiller_SRCS
 		Markers/HeroExchange.cpp
 		Markers/UnlockCluster.cpp
 		Markers/DefendTown.cpp
+		Markers/ExplorationPoint.cpp
 		Engine/Nullkiller.cpp
 		Engine/DeepDecomposer.cpp
 		Engine/PriorityEvaluator.cpp
@@ -57,6 +60,7 @@ set(Nullkiller_SRCS
 		Behaviors/GatherArmyBehavior.cpp
 		Behaviors/ClusterBehavior.cpp
 		Behaviors/StayAtTownBehavior.cpp
+		Behaviors/ExplorationBehavior.cpp
 		Helpers/ArmyFormation.cpp
 		AIGateway.cpp
 )
@@ -80,6 +84,8 @@ set(Nullkiller_HEADERS
 		Pathfinding/Rules/AIMovementToDestinationRule.h
 		Pathfinding/Rules/AIPreviousNodeRule.h
 		Pathfinding/ObjectGraph.h
+		Pathfinding/GraphPaths.h
+		Pathfinding/ObjectGraphCalculator.h
 		AIUtility.h
 		pforeach.h
 		Analyzers/ArmyManager.h
@@ -111,6 +117,7 @@ set(Nullkiller_HEADERS
 		Markers/HeroExchange.h
 		Markers/UnlockCluster.h
 		Markers/DefendTown.h
+		Markers/ExplorationPoint.h
 		Engine/Nullkiller.h
 		Engine/DeepDecomposer.h
 		Engine/PriorityEvaluator.h
@@ -126,6 +133,7 @@ set(Nullkiller_HEADERS
 		Behaviors/GatherArmyBehavior.h
 		Behaviors/ClusterBehavior.h
 		Behaviors/StayAtTownBehavior.h
+		Behaviors/ExplorationBehavior.h
 		Helpers/ArmyFormation.h
 		AIGateway.h
 )
