+++ AI/Nullkiller/Goals/AbstractGoal.h
@@ -73,7 +73,9 @@ namespace Goals
 		CAPTURE_OBJECT,
 		SAVE_RESOURCES,
 		STAY_AT_TOWN_BEHAVIOR,
-		STAY_AT_TOWN
+		STAY_AT_TOWN,
+		EXPLORATION_BEHAVIOR,
+		EXPLORATION_POINT
 	};
 
 	class DLL_EXPORT TSubgoal : public std::shared_ptr<AbstractGoal>
